requirements.txt
asgiref==3.2.10
certifi==2021.10.8
Django==3.1
djangorestframework==3.13.0
djangorestframework-simplejwt==5.0.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
wincertstore==0.2
